move metadata_visualizations to analysis
