+++ gender_analysis/__init__.py
@@ -3,4 +3,3 @@ from .text.corpus import Corpus
 from .text.document import Document
 from .gender.gender import Gender
 from .gender.pronouns import PronounSeries
-from .text.metadata_visualizations import *
