+++ gender_analysis/analysis/__init__.py
@@ -4,8 +4,10 @@ __all__ = [
     'gender_frequency',
     'instance_distance',
     'proximity',
+    'metadata_visualizations',
 ]
 
+from .metadata_visualizations import *
 from .dependency_parsing import *
 from .dunning import *
 from .gender_frequency import *
